requirements.txt
asgiref==3.5.2
attrs==22.1.0
boto3==1.26.36
botocore==1.29.36
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
Faker==15.3.4
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
isodate==0.6.1
jmespath==1.0.1
lxml==4.9.2
packaging==22.0
platformdirs==2.6.0
pluggy==1.0.0
pytest==7.2.0
pytest-django==4.5.2
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
zeep==4.2.1
psycopg2